backend-fastapi/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
beautifulsoup4